Minor fixes (#319)

+++ src/python/utils/delete_keys_from_cache.py
@@ -1,32 +1,17 @@
 # -*- coding: utf-8 -*-
-"""Delete the keys containing 'imports' in the JSON files in the given directory."""
+"""Delete the keys matching a given regex in the JSON files in the given directory."""
 
 import os
 import sys
 import json
+import re
 from pathlib import Path
 from argparse import ArgumentParser
 
-
-def count_import_keys(directory):
-    """Count the number of keys containing 'imports' in the JSON files in the given directory."""
-    count = 0
-    for root, _, files in os.walk(directory):
-        json_files = [f for f in files if f.endswith(".json")]
-        for json_file in json_files:
-            file_path = os.path.join(root, json_file)
-            with open(file_path, "r", encoding="utf-8") as file:
-                data = json.load(file)
-
-            # Count keys containing 'adjacent'
-            keys_to_delete = [key for key in data if "intellimerge" in key]
-            count += len(keys_to_delete)
-    return count
-
-
-def delete_import_keys(directory):
-    """Delete the keys containing 'imports' in the JSON files in the given directory."""
+def delete_keys_matching_regex(directory:Path, regex:str, dry_run:bool=False):
+    """Delete the keys matching the given regex in the JSON files in the given directory."""
     total_deleted = 0
+    pattern = re.compile(regex)
     for root, _, files in os.walk(directory):
         json_files = [f for f in files if f.endswith(".json")]
         for json_file in json_files:
@@ -35,15 +20,16 @@ def delete_import_keys(directory):
                 data = json.load(file)
 
             # Record keys to delete
-            keys_to_delete = [key for key in data.keys() if "intellimerge" in key]
+            keys_to_delete = [key for key in data.keys() if pattern.search(key)]
             if keys_to_delete:
                 for key in keys_to_delete:
                     del data[key]
                     total_deleted += 1
 
-            # Save the modified data back to file
-            with open(file_path, "w", encoding="utf-8") as file:
-                json.dump(data, file, indent=4)
+            if not dry_run:
+                # Save the modified data back to file
+                with open(file_path, "w", encoding="utf-8") as file:
+                    json.dump(data, file, indent=4)
 
     return total_deleted
 
@@ -58,23 +44,31 @@ def main():
         help="The cache directory to delete keys from.",
     )
     parser.add_argument(
+        "--regex",
+        type=str,
+        required=True,
+        help="The regex to match keys for deletion.",
+    )
+    parser.add_argument(
         "-y", "--yes", action="store_true", help="Skip the confirmation prompt."
     )
     args = parser.parse_args()
     cache_dir = Path(args.cache)
+    regex = args.regex
+
     if not cache_dir.exists():
         print(f"Directory '{cache_dir}' does not exist.")
         return
 
     if not args.yes:
-        potential_deletions = count_import_keys(args.cache)
+        potential_deletions = delete_keys_matching_regex(args.cache, regex, dry_run=True)
         print(f"Potential deletions: {potential_deletions}")
         confirm = input("Do you want to proceed with deleting these keys? (yes/no): ")
         if confirm.lower() != "yes":
             print("Operation cancelled.")
             sys.exit(0)
 
-    total_deleted = delete_import_keys(args.cache)
+    total_deleted = delete_keys_matching_regex(args.cache, regex)
     print(f"Total keys deleted: {total_deleted}")
